pyproject.toml
[project]
name = "mcp-x-server"
version = "0.1.0"
description = "Minimalist X (Twitter) MCP server"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "twikit>=2.0.0",
    "twscrape>=0.5.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
